notification-service/requirements.txt
fastapi
uvicorn
httpx==0.28.1
requests
pytest
fastapi[testclient]
python-dotenv
